# lib/api-zod/src/generated/types/createProposalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProposalBody {
  clientName: string;
  projectName: string;
  projectDate: string;
  clientIndustry: string;
  projectType: string;
  budgetRange?: string | null;
  logoUrl?: string | null;
  contactDetails?: string | null;
  signatureUrl?: string | null;
}