# lib/api-zod/src/generated/types/fullGeneratedContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FullGeneratedContent {
  executiveSummary: string;
  aboutCompany: string;
  projectOverview: string;
  features: string;
  technologyStack: string;
  pricing: string;
  digitalMarketing: string;
  addOns: string;
  legalTerms: string;
  acceptanceSection: string;
}