# lib/api-zod/src/generated/types/generateFullProposalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateFullProposalBody {
  projectType: string;
  clientIndustry: string;
  clientName: string;
  projectName: string;
  budgetRange?: string | null;
}