# lib/api-zod/src/generated/types/proposal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProposalEnabledSections } from "./proposalEnabledSections";
import type { ProposalSections } from "./proposalSections";

export interface Proposal {
  id: number;
  clientName: string;
  projectName: string;
  projectDate: string;
  clientIndustry: string;
  projectType: string;
  budgetRange?: string | null;
  logoUrl?: string | null;
  contactDetails?: string | null;
  signatureUrl?: string | null;
  status: string;
  sections?: ProposalSections;
  enabledSections?: ProposalEnabledSections;
  createdAt: string;
  updatedAt: string;
}