# lib/api-zod/src/generated/types/proposalStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Proposal } from "./proposal";

export interface ProposalStats {
  total: number;
  draft: number;
  sent: number;
  approved: number;
  recentProposals: Proposal[];
}