# lib/api-zod/src/generated/types/updateProposalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProposalBodyEnabledSections } from "./updateProposalBodyEnabledSections";
import type { UpdateProposalBodySections } from "./updateProposalBodySections";

export interface UpdateProposalBody {
  clientName?: string;
  projectName?: string;
  projectDate?: string;
  clientIndustry?: string;
  projectType?: string;
  budgetRange?: string | null;
  logoUrl?: string | null;
  contactDetails?: string | null;
  signatureUrl?: string | null;
  status?: string;
  sections?: UpdateProposalBodySections;
  enabledSections?: UpdateProposalBodyEnabledSections;
}