# lib/api-zod/src/generated/types/uploadLogoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadLogoBody {
  base64Data: string;
  mimeType: string;
  fileName: string;
}